woodtick-pipeline/requirements.txt
# woodtick-pipeline requirements
# Python 3.12 / Airflow 2.8 ETL data pipeline

apache-airflow==2.8.1
celery==5.3.6
redis==5.0.1
pandas==2.1.4
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
minio==7.2.3
boto3==1.34.14
pyyaml==6.0.1
# BUG-0001: Jinja2 pinned to old version with known SSTI vulns (CWE-1395, CVSS 4.0, BEST_PRACTICE, Tier 5)
Jinja2==3.1.2
cryptography==41.0.7
requests==2.31.0
python-dotenv==1.0.0
flask==3.0.0
gunicorn==21.2.0
marshmallow==3.20.2
# BUG-0002: pickle5 enables unsafe deserialization patterns (CWE-502, CVSS 5.0, TRICKY, Tier 6)
pickle5==0.0.12; python_version < "3.8"
numpy==1.26.3
openpyxl==3.1.2
xlrd==2.0.1
